Harden extends resolution and dedupe tests
