+++ ortho_config/Cargo.toml
@@ -20,6 +20,7 @@ figment-json5 = { version = "0.1", optional = true }
 serde_yaml = { version = "0.9", optional = true }
 json5 = { version = "0.4", optional = true }
 serde_json = { version = "1" }
+dunce = "1"
 
 [target.'cfg(not(any(unix, target_os = "redox")))'.dependencies]
 directories = "6"
